+++ cmd/npmi-go/flags.go
@@ -1,4 +1,4 @@
-package cli
+package main
 
 import (
 	"flag"
@@ -7,32 +7,8 @@ import (
 	"os"
 
 	"github.com/caarlos0/env/v6"
+	"github.com/hermo/npmi-go/pkg/npmi"
 )
-
-// MinioCacheOptions contains configuration for Minio Cache
-type MinioCacheOptions struct {
-	Endpoint        string `env:"NPMI_MINIO_ENDPOINT"`
-	AccessKeyID     string `env:"NPMI_MINIO_ACCESS_KEY_ID"`
-	SecretAccessKey string `env:"NPMI_MINIO_SECRET_ACCESS_KEY"`
-	Bucket          string `env:"NPMI_MINIO_BUCKET"`
-	UseTLS          bool   `env:"NPMI_MINIO_TLS"`
-}
-
-// LocalCacheOptions constains configuration for Local Cache
-type LocalCacheOptions struct {
-	Dir string `env:"NPMI_LOCAL_DIR"`
-}
-
-// Options describes the runtime configuration
-type Options struct {
-	Verbose         bool `env:"NPMI_VERBOSE"`
-	Force           bool `env:"NPMI_FORCE"`
-	UseLocalCache   bool `env:"NPMI_LOCAL"`
-	UseMinioCache   bool `env:"NPMI_MINIO"`
-	MinioCache      *MinioCacheOptions
-	LocalCache      *LocalCacheOptions
-	PrecacheCommand string `env:"NPMI_PRECACHE"`
-}
 
 var (
 	version = "dev"
@@ -77,18 +53,18 @@ OPTIONS:
 )
 
 // ParseFlags parses command line flags
-func ParseFlags() (*Options, error) {
-	options := &Options{
+func ParseFlags() (*npmi.Options, error) {
+	options := &npmi.Options{
 		Verbose:         false,
 		Force:           false,
 		UseLocalCache:   true,
 		UseMinioCache:   false,
 		PrecacheCommand: "",
 	}
-	localCache := &LocalCacheOptions{
+	localCache := &npmi.LocalCacheOptions{
 		Dir: os.TempDir(),
 	}
-	minioCache := &MinioCacheOptions{
+	minioCache := &npmi.MinioCacheOptions{
 		Endpoint:        "",
 		AccessKeyID:     "",
 		SecretAccessKey: "",
