+++ pkg/npmi/installer_test.go
@@ -2,11 +2,17 @@ package npmi
 
 import (
 	"testing"
+
+	"github.com/hermo/npmi-go/pkg/cmd"
 )
 
 func TestInstaller_Run(t *testing.T) {
 
-	runner := &SpyRunner{"v11.16.3-darwin-x64", "", nil}
+	runner := &cmd.SpyRunner{
+		Stdout: "v11.16.3-darwin-x64",
+		Stderr: "",
+		Error:  nil,
+	}
 	nc := &Config{
 		runner:         runner,
 		productionMode: false,
